ref: move sigint handler out of fetcher
