+++ state-reconstruct-fetcher/Cargo.toml
@@ -6,20 +6,20 @@ edition = "2021"
 [dependencies]
 bincode = "1.3.3"
 blake2 = "0.10.6"
+chrono = "0.4.31"
 ethers = "1.0.2"
 eyre = "0.6.8"
+hex = "0.4.3"
 indexmap = { version = "2.0.2", features = ["serde"] }
 rand = "0.8.5"
 reqwest = "0.11.24"
+rocksdb = "0.21.0"
 serde = { version = "1.0.189", features = ["derive"] }
 serde_json = { version = "1.0.107", features = ["std"] }
 serde_json_any_key = "2.0.0"
+state-reconstruct-storage = { path = "../state-reconstruct-storage" }
 thiserror = "1.0.50"
-tokio = { version = "1.33.0", features = ["signal"] }
+tokio = "1.33.0"
 tokio-util = "0.7.10"
 tracing = "0.1.40"
-rocksdb = "0.21.0"
-hex = "0.4.3"
-chrono = "0.4.31"
 zkevm_circuits = { git = "https://github.com/matter-labs/era-zkevm_circuits", branch = "v1.4.1" }
-state-reconstruct-storage = { path = "../state-reconstruct-storage" }
